+++ blacksheep/server/application.py
@@ -64,7 +64,10 @@ from blacksheep.server.routing import (
     RoutesRegistry,
 )
 from blacksheep.server.routing import router as default_router
-from blacksheep.server.routing import validate_default_router, validate_router
+from blacksheep.server.routing import (
+    validate_default_router,
+    validate_router,
+)
 from blacksheep.server.websocket import WebSocket, format_reason
 from blacksheep.sessions import SessionMiddleware, SessionSerializer
 from blacksheep.settings.di import di_settings