+++ blacksheep/server/openapi/ui.py
@@ -3,11 +3,10 @@ from dataclasses import dataclass
 from typing import Callable, Optional
 
 from blacksheep.messages import Request, Response
-from blacksheep.server.env import get_global_route_prefix
 from blacksheep.server.files.static import get_response_for_static_content
 from blacksheep.server.resources import get_resource_file_content
+from blacksheep.server.responses import moved_permanently
 from blacksheep.utils.time import utcnow
-from blacksheep.url import join_prefix
 
 SWAGGER_UI_JS_URL = (
     "https://cdn.jsdelivr.net/npm/swagger-ui-dist@5/swagger-ui-bundle.js"
@@ -34,15 +33,6 @@ class UIOptions:
     spec_url: str
     page_title: str
     favicon_url: str = "favicon.png"
-
-    def __post_init__(self):
-        # If a global route prefix is defined, it is applied to the spec URL and
-        # the favicon URL.
-        global_prefix = get_global_route_prefix()
-
-        if global_prefix:
-            self.spec_url = join_prefix(global_prefix, self.spec_url)
-            self.favicon_url = join_prefix(global_prefix, self.favicon_url)
 
 
 class UIProvider(ABC):
@@ -104,6 +94,12 @@ class SwaggerUIProvider(UIProvider):
         current_time = utcnow().timestamp()
 
         def get_open_api_ui(request: Request) -> Response:
+            path = request.path
+
+            # This is needed to ensure relative URLs in the Swagger UI work properly.
+            if not path.endswith("/"):
+                return moved_permanently(f"/{path.strip('/')}/")
+
             return get_response_for_static_content(
                 request, b"text/html; charset=utf-8", self._ui_html, current_time
             )
@@ -142,6 +138,12 @@ class ReDocUIProvider(UIProvider):
         current_time = utcnow().timestamp()
 
         def get_open_api_ui(request: Request) -> Response:
+            path = request.path
+
+            # This is needed to ensure relative URLs in the Swagger UI work properly.
+            if not path.endswith("/"):
+                return moved_permanently(f"/{path.strip('/')}/")
+
             return get_response_for_static_content(
                 request, b"text/html; charset=utf-8", self._ui_html, current_time
             )