+++ tests/client/test_headers.py
@@ -1,5 +1,3 @@
-import pytest
-
 from blacksheep import Response, TextContent
 from blacksheep.client import ClientSession