+++ tests/test_auth_cookie.py
@@ -1,6 +1,5 @@
 from typing import Any
 
-import pytest
 from guardpost import Identity
 
 from blacksheep.messages import Request, Response